requirements.txt
appnope==0.1.0
astroid==2.2.5
atomicwrites==1.3.0
attrs==19.1.0
autopep8==1.4.4
backcall==0.1.0
bleach==3.1.0
certifi==2019.3.9
chardet==3.0.4
Click==7.0
click-plugins==1.1.1
cligj==0.5.0
colorlog==4.0.2
coverage==4.5.3
cycler==0.10.0
decorator==4.4.0
defusedxml==0.6.0
descartes==1.1.0
dj-database-url==0.5.0
Django==2.2.1
django-extensions==2.1.7
djangorestframework==3.9.4
entrypoints==0.3
Fiona==1.8.6
GDAL==2.4.1
geopandas==0.5.0
idna==2.8
ipykernel==5.1.1
ipython==7.5.0
ipython-genutils==0.2.0
ipywidgets==7.4.2
isort==4.3.20
jedi==0.13.3
Jinja2==2.10.1
jsonschema==3.0.1
jupyter==1.0.0
jupyter-client==5.2.4
jupyter-console==6.0.0
jupyter-contrib-core==0.3.3
jupyter-contrib-nbextensions==0.5.1
jupyter-core==4.4.0
jupyter-highlight-selected-word==0.2.0
jupyter-latex-envs==1.4.6
jupyter-nbextensions-configurator==0.4.1
kiwisolver==1.1.0
lazy-object-proxy==1.4.1
lxml==4.3.3
mapclassify==2.0.1
MarkupSafe==1.1.1
matplotlib==3.1.0
mccabe==0.6.1
mistune==0.8.4
more-itertools==7.0.0
munch==2.3.2
nbconvert==5.5.0
nbformat==4.4.0
notebook==5.7.8
numpy==1.16.3
palettable==3.1.1
pandas==0.24.2
pandoc==1.0.2
pandocfilters==1.4.2
parso==0.4.0
pexpect==4.7.0
pickleshare==0.7.5
plotly==3.9.0
pluggy==0.11.0
ply==3.11
prometheus-client==0.6.0
prompt-toolkit==2.0.9
protobuf==3.7.1
psycopg2==2.8.2
ptyprocess==0.6.0
py==1.8.0
pycodestyle==2.5.0
Pygments==2.4.1
pylint==2.3.1
pylint-django==2.0.9
pylint-plugin-utils==0.5
pyparsing==2.4.0
pyproj==2.1.3
pyrsistent==0.15.2
pysal==2.0.0
pytest==4.5.0
pytest-cov==2.7.1
pytest-django==3.4.8
python-dateutil==2.8.0
python-dotenv==0.10.2
pytz==2019.1
PyYAML==5.1
pyzmq==18.0.1
qtconsole==4.5.0
requests==2.22.0
retrying==1.3.3
rope==0.14.0
scipy==1.3.0
seaborn==0.9.0
Send2Trash==1.5.0
Shapely==1.6.4.post2
six==1.12.0
sqlparse==0.3.0
terminado==0.8.2
testpath==0.4.2
tornado==6.0.2
traitlets==4.3.2
typed-ast==1.3.5
urllib3==1.25.3
virtualenv==16.6.0
wcwidth==0.1.7
webencodings==0.5.1
widgetsnbextension==3.4.2
wrapt==1.11.1
